requirements.txt
asgiref==3.6.0
boto3==1.26.142
botocore==1.29.142
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-storages==1.13.2
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
Pillow==9.5.0
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
urllib3==1.26.16
